修改uid全为Long

+++ webserver/truemen-api-user/src/main/java/com/truemen/api/user/dao/UserDAO.java
@@ -108,7 +108,7 @@ public class UserDAO {
         }
     }
 
-    public User getUserByUid(int uid) {
+    public User getUserByUid(Long uid) {
         String sql = "SELECT * FROM userCoreInfo uci JOIN userBaseInfo ubi ON uci.uid = ubi.uid WHERE uci.uid = ?";
         try {
             User user = jdbcTemplate.queryForObject(sql, userRowMapper, uid);
